Restructure to use separate statement visitor
